requirements.txt
aiohttp==3.7.2
alpha-vantage==2.2.0
argon2-cffi==20.1.0
asgiref==3.2.10
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.2.1
Automat==20.2.0
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.5
cssutils==1.0.2
daphne==3.0.1
decorator==4.4.2
defusedxml==0.6.0
Faker==11.1.0
factory_boy==3.2.1
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.6.0
django-environ==0.4.5
django-extensions==3.0.9
django-inlinecss==0.3.0
django-phonenumber-field==1.1.0
django-rest-passwordreset==1.1.0
django-sendgrid-v5==0.8.1
django-taggit==1.3.0
django-filter==21.1
django-tree-queries==0.6.0
djangorestframework==3.11.0
drf-yasg==1.20.0
entrypoints==0.3
future==0.18.2
gunicorn==20.0.4
hyperlink==21.0.0
idna==2.8
incremental==17.5.0
inflection==0.5.1
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
MarkupSafe==1.1.1
mistune==0.8.4
multidict==5.0.0
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.1.4
numpy==1.19.5
packaging==20.8
pandas==1.0.3
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
phonenumbers==8.12.4
phonenumberslite==8.12.11
pickleshare==0.7.5
Pillow==7.1.2
prometheus-client==0.8.0
prompt-toolkit==3.0.16
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.0
PyHamcrest==2.0.2
pynliner==0.8.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-decouple==3.3
python-http-client==3.3.1
pytz==2020.1
pyzmq==19.0.2
qtconsole==4.7.7
python-dotenv==0.19.0
QtPy==1.9.0
requests==2.21.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
Send2Trash==1.5.0
sendgrid==6.4.7
service-identity==18.1.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
terminado==0.9.1
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
Twisted==20.3.0
txaio==20.12.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.24.3
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
yarl==1.6.2
zope.interface==5.2.0
channels==3.0.4
channels_redis==3.4.0